Module Minicaml.Optimizer

val optimize : Types.expr -> Types.expr

val optimize_let : (Types.ide * Types.expr) list -> Types.expr -> bool -> Types.expr

val iterate_optimizer : Types.expr -> Types.expr

Apply the optimizer again and again on an expression until it is fully reduced and ready to be evaluated