Module Minicaml.Types

type ide = string

A value identifier

val pp_ide : Ppx_deriving_runtime.Format.formatter -> ide -> Ppx_deriving_runtime.unit

val show_ide : ide -> Ppx_deriving_runtime.string

val equal_ide : ide -> ide -> Ppx_deriving_runtime.bool

val compare_ide : ide -> ide -> Ppx_deriving_runtime.int

type complext = Stdlib.Complex.t

A type wrapper for complex numbers where equality, ordering and showing are defined

val pp_complext : Ppx_deriving_runtime.Format.formatter -> complext -> Ppx_deriving_runtime.unit

val show_complext : complext -> Ppx_deriving_runtime.string

val equal_complext : complext -> complext -> Ppx_deriving_runtime.bool

val compare_complext : complext -> complext -> Ppx_deriving_runtime.int

type puret =

| Uncertain

| Pure

| Impure

A type representing if a computation is pure or not

val pp_puret : Ppx_deriving_runtime.Format.formatter -> puret -> Ppx_deriving_runtime.unit

val show_puret : puret -> Ppx_deriving_runtime.string

val equal_puret : puret -> puret -> Ppx_deriving_runtime.bool

val compare_puret : puret -> puret -> Ppx_deriving_runtime.int

type expr =

| Unit

| Purity of puret * expr

| NumInt of int

| NumFloat of float

| NumComplex of complext

| Boolean of bool

| String of string

| Symbol of ide

| List of expr list

| Cons of expr * expr

| ConcatLists of expr * expr

| ConcatStrings of expr * expr

| Dict of (expr * expr) list

| Plus of expr * expr

| Sub of expr * expr

| Div of expr * expr

| Mult of expr * expr

| Eq of expr * expr

| Gt of expr * expr

| Lt of expr * expr

| Ge of expr * expr

| Le of expr * expr

| And of expr * expr

| Or of expr * expr

| Not of expr

| IfThenElse of expr * expr * expr

| Let of (ide * expr) list * expr

| Letlazy of (ide * expr) list * expr

| Letrec of ide * expr * expr

| Letreclazy of ide * expr * expr

| Lambda of ide list * expr

| Apply of expr * expr list

| Sequence of expr list

| Pipe of expr * expr

The type representing Abstract Syntax Tree expressions

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

val equal_expr : expr -> expr -> Ppx_deriving_runtime.bool

val compare_expr : expr -> expr -> Ppx_deriving_runtime.int

type command =

| Expr of expr

| Def of (ide * expr) list

| Defrec of (ide * expr) list

A type useful for evaluating files, stating if a command is an expression or simply a "global" declaration (appended to environment)

val pp_command : Ppx_deriving_runtime.Format.formatter -> command -> Ppx_deriving_runtime.unit

val show_command : command -> Ppx_deriving_runtime.string

val equal_command : command -> command -> Ppx_deriving_runtime.bool

val compare_command : command -> command -> Ppx_deriving_runtime.int

type 'a env_t = (string * 'a) list

A purely functional environment type, parametrized

val pp_env_t : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a env_t -> Ppx_deriving_runtime.unit

val show_env_t : a. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a env_t -> Ppx_deriving_runtime.string

val equal_env_t : a. ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a env_t -> 'a env_t -> Ppx_deriving_runtime.bool

val compare_env_t : a. ('a -> 'a -> Ppx_deriving_runtime.int) -> 'a env_t -> 'a env_t -> Ppx_deriving_runtime.int

type evt =

| EvtUnit

| EvtInt of int

| EvtFloat of float

| EvtComplex of complext

| EvtBool of bool

| EvtString of string

| EvtList of evt list

| EvtDict of (evt * evt) list

| Closure of ide list * expr * type_wrapper env_t
RecClosure keeps the function name in the constructor for recursion

| RecClosure of ide * ide list * expr * type_wrapper env_t
Abstraction that permits treating primitives as closures

| PrimitiveAbstraction of primitivet

A type that represents an evaluated (reduced) value

and type_wrapper =

| LazyExpression of expr

| AlreadyEvaluated of evt

and primitivet = ide * int * type_wrapper env_t * puret

val pp_evt : Ppx_deriving_runtime.Format.formatter -> evt -> Ppx_deriving_runtime.unit

val show_evt : evt -> Ppx_deriving_runtime.string

val pp_type_wrapper : Ppx_deriving_runtime.Format.formatter -> type_wrapper -> Ppx_deriving_runtime.unit

val show_type_wrapper : type_wrapper -> Ppx_deriving_runtime.string

val pp_primitivet : Ppx_deriving_runtime.Format.formatter -> primitivet -> Ppx_deriving_runtime.unit

val show_primitivet : primitivet -> Ppx_deriving_runtime.string

val equal_evt : evt -> evt -> Ppx_deriving_runtime.bool

val equal_type_wrapper : type_wrapper -> type_wrapper -> Ppx_deriving_runtime.bool

val equal_primitivet : primitivet -> primitivet -> Ppx_deriving_runtime.bool

val compare_evt : evt -> evt -> Ppx_deriving_runtime.int

val compare_type_wrapper : type_wrapper -> type_wrapper -> Ppx_deriving_runtime.int

val compare_primitivet : primitivet -> primitivet -> Ppx_deriving_runtime.int

val generate_prim_params : int -> string list

val show_unpacked_evt : evt -> Ppx_deriving_runtime.string

type env_type = type_wrapper env_t

An environment of already evaluated values

type stackframe =

| StackValue of int * expr * stackframe

| EmptyStack

A recursive type representing a stacktrace frame

val pp_stackframe : Ppx_deriving_runtime.Format.formatter -> stackframe -> Ppx_deriving_runtime.unit

val show_stackframe : stackframe -> Ppx_deriving_runtime.string

val push_stack : stackframe -> expr -> stackframe

Push an AST expression into a stack

parameter s

The stack where to push the expression

parameter e

The expression to push

val pop_stack : stackframe -> stackframe

Pop an AST expression from a stack

val depth_of_stack : stackframe -> int

type evalstate = {

env : env_type;

verbosity : int;

stack : stackframe;

printresult : bool;

purity : puret;

}

Options for the eval function, includes

exception UnboundVariable of string

exception WrongPrimitiveArgs

exception TypeError of string

exception ListError of string

exception DictError of string

exception SyntaxError of string

exception FileNotFoundError of string

exception PurityError of string