Module Minicaml.Util

module Dict : sig ... end

val take : int -> 'a list -> 'a list

Helper function to take the first elements of a list

val drop : int -> 'a list -> 'a list

Helper function to drop the first elements of a list

val last : 'a list -> 'a

val fst : ('a * 'b) -> 'a

val snd : ('a * 'b) -> 'b

val fstl : ('a * 'b) list -> 'a list

val sndl : ('a * 'b) list -> 'b list

val unzip : ('a * 'b) list -> 'a list * 'b list

Helper function to unzip a list of couples

val zip : 'a list -> 'b list -> ('a * 'b) list