Module Minicaml__Mstdlib

val parser : Stdlib.Lexing.lexbuf -> Minicaml.Types.command

val closurize : string -> string -> Minicaml.Types.type_wrapper

An helper function that helps extracting closures from strings, to be used as functions in the standard library. An empty environment is used since primitives in the standard library should not be able to access external values TODO: compute at compile time

val mapstr : string

val filterstr : string

val foldlstr : string

val foldrstr : string

val table : (string * Minicaml.Types.type_wrapper) list

val purity_table : (string * Minicaml.Types.puret) list