Module Minicaml__Optimizer

val optimize : Minicaml.Types.expr -> Minicaml.Types.expr

val optimize_let : Minicaml.Types.assignment_type list -> Minicaml.Types.expr -> Minicaml.Types.expr

val iterate_optimizer : Minicaml.Types.expr -> Minicaml.Types.expr

Apply the optimizer again and again on an expression until it is fully reduced and ready to be evaluated