Module Minicaml__Puritycheck

val level_purity : Minicaml.Types.puret -> Minicaml.Types.puret -> Minicaml.Types.puret

"level out" the purity of two values

val infer : Minicaml.Types.expr -> Minicaml.Types.evalstate -> Minicaml.Types.puret

Infer the purity of an expression. Note: this is a naive approach. This function is an abstract interpretation of expressions over primitives and environments.

parameter pt

The primitives table

parameter paraml

If inside a lambda, the list of parameters

parameter env

An environment, needed to infer whether lazy expressions are pure or not

val lookup : Minicaml.Types.ide -> Minicaml.Types.evalstate -> Minicaml.Types.puret

val infer_assignment : Minicaml.Types.evalstate -> (bool * Minicaml.Types.ide * Minicaml.Types.expr) -> Minicaml.Types.evalstate

val infer_assignment_list : Minicaml.Types.assignment_type list -> Minicaml.Types.evalstate -> Minicaml.Types.evalstate