Module Minicaml__Util.Dict

val dup_exists : 'a list -> bool

Search for duplicates

val empty : unit -> Minicaml.Types.env_type

Generate an empty dictionary

val delete : 'a -> ('a * 'b) list -> ('a * 'b) list

Delete all occurrences of a key in a dictionary

val insert : ('a * 'b) list -> 'a -> 'b -> ('a * 'b) list

Insert a value into a key-value dictionary

val exists : 'a -> ('a * 'b) list -> bool

Check if a key exists in a list of key-value pairs

val get : 'a -> ('a * 'b) list -> 'b

Search and Dict.get a key's value (first match) in a list of key-value pairs

val filter : 'a list -> ('a * 'b) list -> ('a * 'b) list

val insertmany : ('a * 'b) list -> 'a list -> 'b list -> ('a * 'b) list

Insert a list of keys and a list of values in a dictionary